+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 # FollowersCompare Changelog
 
 *Last updated: 2026-06-08*  
-*Current Version: v0.2.2*
+*Current Version: v0.2.5*
 
 ## Overview
-This document tracks past changes starting with v0.2.3
+This document tracks past changes starting with v0.2.5
 
 ## Changelog
 
-### Release v0.2.3
+### Release v0.2.5
+- Improved Uploading Handling
+- Structure Updates
+- Added the ability to hide deactivated and deleted accounts